add new feature (add money to balance using the credit card #55

+++ src/main/java/com/example/wayz/Model/Student.java
@@ -38,7 +38,7 @@ public class Student {
     private Integer tripsLeft=0;
 
     @Column(nullable = false)
-    private Double balance=0.0;
+    private Integer balance=0;
 
 
     @OneToOne
